requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
aws-cdk-lib
aws-cdk.aws-lambda-python-alpha
fastapi
mangum
tai-aws-account-bootstrap
